pyproject.toml
[tool.poetry]
name = "zlm"
version = "0.1.35"
description = "Simplify and improve your job hunting experience by integrating LLMs to automate tasks such as resume and cover letter generation, as well as application submission, saving users time and effort."
authors = ["Saurabh Zinjad <zinjadsaurabh1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.6,<3.13"
streamlit = "^1.37.1"
validators = "^0.33.0"
langchain-core = "^0.2.33"
langchain-community = "^0.2.12"
pdf2image = "^1.17.0"
openai = "^1.41.0"
google-generativeai = "^0.7.2"
pypdf2 = "^3.0.1"
beautifulsoup4 = "^4.12.3"
scikit-learn = "^1.5.1"
nltk = "^3.8.1"
langchain-ollama = "^0.1.1"
playwright = "^1.46.0"
unstructured = "^0.15.5"
markdown-pdf = "^1.3"
fpdf2 = "^2.8.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"